Load ArchiveManager on launch if possible
