+++ WolvenKit.App/Controllers/IGameController.cs
@@ -76,8 +76,6 @@ public interface IGameController
     /// <returns>bool success</returns>
     Task<bool> AddFileToModModalAsync(ulong hash, ArchiveManagerScope searchScope);
 
-    public Task HandleStartup();
-
     Task<bool> LaunchProjectAsync(LaunchProfile profile);
     Task<bool> InstallProjectHotAsync();
     bool CleanAll(bool isPostBuild = false);
