+++ WolvenKit.App/Services/IArchiveManagerLoader.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace WolvenKit.App.Services;
+
+public interface IArchiveManagerLoader
+{
+    const string ArchiveLoadingPurpose = "RED4Controller archive loading";
+
+    Task LoadArchiveManagerAsync();
+}
